adk_hello_streamlit/pyproject.toml
[project]
name = "adk-hello-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.17.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
